+++ model/data.py
@@ -6,22 +6,20 @@ import numpy as np
 from arch.unitroot import PhillipsPerron
 from statsmodels.tsa.stattools import adfuller
 from datetime import date
-from scipy.signal import butter, filtfilt
+
+from features.standart_features import StandartFeaturesMixin
+from features.time_features import TimeFeaturesMixin
+from features.trend_features import TrendFeaturesMixin
+from features.uncommon_features import UncommonFeaturesMixin
 
 # Здесь все признаки и все по датафрейму
 
 # to do:
 # Добавить все признаки, которые можно сюда добавить 
 
-def butter_filter(data, cutoff_frequency, fs, order=4):
-    nyquist = 0.5 * fs
-    normal_cutoff = cutoff_frequency / nyquist
-    b, a = butter(order, normal_cutoff, btype='low', analog=False)
-    return filtfilt(b, a, data)
 
 
-
-class FinData():
+class FinData(StandartFeaturesMixin, TimeFeaturesMixin, TrendFeaturesMixin, UncommonFeaturesMixin):
     """
     Класс для обработки финансовых данных. 
     Позволяет загружать данные, фильтровать их по времени, добавлять признаки, 
@@ -197,170 +195,6 @@ class FinData():
 
         plt.legend(facecolor='lightgrey', edgecolor='black', title='Columns')
 
-    
-    # Добавление признаков 
-    def insert_shifts_norms(self, windows_shifts_norms = [3, 6, 18]):
-        """
-        Добавляет нормализованные значения цены с учетом сдвигов.
-
-        Параметры:
-            windows_shifts_norms (list): Список сдвигов для нормализации.
-        """
-        for i in windows_shifts_norms:
-            self.df[f'close_norms_{i}'] = self.df['close']/self.df['close'].shift(i)
-            self.df[f'close_high_norms_{i}'] = self.df['close']/self.df['close'].shift(i)
-            self.df[f'high_norms_{i}'] = self.df['high']/self.df['high'].shift(i)
-            self.df[f'low_norms_{i}'] = self.df['low']/self.df['low'].shift(i)
-
-            if f'close_norms_{i}' not in self.numeric_features:
-                self.numeric_features += [f'close_norms_{i}', f'close_high_norms_{i}', f'high_norms_{i}', f'low_norms_{i}']
-
-    def insert_time_features(self):
-        # Добавляет минуты, дни, часы 
-        # Убрать может раз вообще незначимо
-                                  
-        self.df['hours'] = self.df['utc'].dt.hour
-        self.df['day'] = self.df['utc'].dt.day_of_year
-        self.df['minute'] = (self.df['utc'].dt.minute + 60 * self.df['hours'])
-
-        if 'hours' not in self.cat_features:
-            self.cat_features += ['hours', 'day', 'minute']
-    
-    def insert_holidays(self):
-    
-        self.df['is_holiday'] = self.df['utc'].dt.strftime('%m-%d').isin(['12-31', '01-01', '01-07', '02-14', '02-23', '03-08', '05-01']).astype('int')
-        self.df['pred_holiday'] = self.df['utc'].dt.strftime('%m-%d').isin(['11-11', '11-30', '12-30', '02-13', '02-22', '03-07', '04-30']).astype('int') # + black friday
-        self.df['week_pred_holiday'] = self.df['utc'].dt.strftime('%m-%d').isin(['12-24', '02-07', '02-16', '03-01', '04-24']).astype('int')
-
-        if 'is_holiday' not in self.cat_features:
-            self.cat_features += ['is_holiday', 'pred_holiday', 'week_pred_holiday']
-
-
-    def insert_rolling_means(self, windows_ma = [3, 6, 18]):
-        """
-        Добавляет скользящие средние для указанных окон.
-
-        Параметры:
-            windows_ma (list): Список размеров окон для скользящих средних.
-        """
-        # скользящие средние 
-        for i in windows_ma:
-            self.df[f'ma_{i}'] = self.df['close'].rolling(window = i, closed="left").mean()
-            self.df[f'close_normed_ma_{i}'] = self.df['close']/self.df[f'ma_{i}']
-
-            if f'ma_{i}' not in self.numeric_features:
-                self.numeric_features += [f'ma_{i}', f'close_normed_ma_{i}']
-            
-    def insert_exp_rolling_means(self, windows_ema = [3, 6, 18]):
-        """
-        Добавляет экспоненциальные скользящие средние для указанных окон.
-
-        Параметры:
-            windows_ema (list): Список размеров окон для EMA.
-        """
-        # экспоненциальные скользящие средние
-        for i in windows_ema:
-            self.df[f'ema_{i}'] = (self.df['close']).ewm(span=i).mean()
-            self.df[f'close_normed_ema_{i}'] = self.df['close']/self.df[f'ema_{i}']
-
-            if f'ema_{i}' not in self.numeric_features:
-                self.numeric_features += [f'ema_{i}', f'close_normed_ema_{i}']
-    
-    def insert_rsi(self, windows_rsi = [3, 6, 18]):
-        for i in windows_rsi:
-            self.df[f'rsi_{i}'] = self.df['close'] - self.df['close'].shift(i)
-            self.df[f'close_normed_rsi_{i}'] = self.df['close']/self.df[f'rsi_{i}']
-
-            if f'rsi_{i}' not in self.numeric_features:
-                self.numeric_features += [f'rsi_{i}', f'close_normed_rsi_{i}']
-    
-    def insert_bollinger(self): # TODO add other windows
-        if 'ma_16' not in self.df.columns:
-            self.insert_rolling_means([16])
-        self.df['SD'] =np.sqrt(np.square((self.df["ma_16"] - self.df["close"])).rolling(16, closed='left').sum())
-        self.df['upper_bollinger'] = self.df["ma_16"] + self.df['SD'] * 2
-        self.df['low_bollinger'] = self.df["ma_18"] - self.df['SD'] * 2
-
-        self.df['close_normed_upper_bollinger'] = self.df['close']/self.df['upper_bollinger']
-        self.df['close_normed_low_bollinger'] = self.df['close']/self.df['low_bollinger']
-
-        if 'SD' not in self.numeric_features:
-                self.numeric_features += ['SD', 'low_bollinger', 'upper_bollinger', 'close_normed_upper_bollinger', 'close_normed_low_bollinger']
-    
-    def insert_high_low_diff(self, windows_hl = [3, 6, 18]):
-        for i in windows_hl:
-            self.df[f'hl_diff_{i}'] = self.df['high'].shift(i) - self.df['low'].shift(i)
-            self.df[f'close_normed_hl_diff_{i}'] = self.df['close']/self.df[f'hl_diff_{i}']
-
-            if f'hl_diff_{i}' not in self.numeric_features:
-                self.numeric_features += [f'hl_diff_{i}', f'close_normed_hl_diff_{i}']
-    
-    def insert_stochastic_oscillator(self, windows_st = [3, 6, 18]): # TODO add other windows
-
-        self.df['stochastic_oscillator'] = (self.df['close'] - self.df['low'].shift(18)) / (self.df['high'].shift(18) - self.df['low'].shift(18))
-        self.df['close_normed_stochastic_oscillator'] = self.df['close'] / self.df['stochastic_oscillator']
-        
-        if 'stochastic_oscillator' not in self.numeric_features:
-            self.numeric_features += ['stochastic_oscillator', 'close_normed_stochastic_oscillator']
-
-        for i in windows_st:
-            self.df[f'stochastic_oscillator_ma_{i}'] = self.df['stochastic_oscillator'].rolling(i).mean()
-            self.df[f'close_normed_stochastic_oscillator_ma_{i}'] = self.df['close'] / self.df[f'stochastic_oscillator_ma_{i}']
-
-            if f'stochastic_oscillator_ma_{i}' not in self.numeric_features:
-                self.numeric_features += [f'stochastic_oscillator_ma_{i}', f'close_normed_stochastic_oscillator_ma_{i}']
-    
-    def insert_random_prediction(self): # TODO think, if other normalization is needed
-        self.df['target_predict'] = np.exp(np.log1p(self.df['close'].shift(15)) + 4 * np.random.normal(1)) - self.df['close']
-        if 'target_predict' not in self.numeric_features:
-            self.numeric_features += ['target_predict']
-    
-    def insert_butter_filter(self, fs=144, cutoff_frequency=5, order=4, visualise=False):
-            self.df['butter_filter_trend'] = butter_filter(self.df['close'].values, cutoff_frequency, fs, order)
-            self.df['butter_filter_std'] = self.df['close'] - self.df['butter_filter_trend']
-            if 'butter_filter_trend' not in self.numeric_features:
-                self.numeric_features += ['butter_filter_trend', 'butter_filter_std']
-
-            if visualise:
-                plt.plot(self.df['utc'], self.df['open'], label='Original')
-                plt.plot(self.df['utc'], self.df['butter_filter_trend'], label='Filtered Trend', color='red')
-                plt.plot(self.df['utc'], self.df['butter_filter_std'], label='Short Term Component', color='blue')
-                plt.xticks(self.df['utc'][::self.df.shape[0] // 10], rotation=45)
-                plt.legend()
-                plt.show()
-    
-    def insert_trend_rsi(self, windows = [3, 6, 18], filters = ['butter']): # TODO generalize for several filters
-        if 'butter_filter_trend' not in self.df.columns:
-            self.insert_butter_filter()
-        
-        for i in windows:
-            self.df[f'butter_filter_trend_rsi_{i}'] = self.df['butter_filter_trend'] - self.df['butter_filter_trend'].shift(i)
-            self.df[f'close_normed_butter_filter_trend_rsi_{i}'] = self.df['close'] / self.df[f'butter_filter_trend_rsi_{i}']
-
-            if f'butter_filter_trend_rsi_{i}' not in self.numeric_features:
-                self.numeric_features += [f'butter_filter_trend_rsi_{i}', f'close_normed_butter_filter_trend_rsi_{i}']
-
-    def insert_trend_rolling_means(self, windows = [3, 6, 18]):
-        if 'butter_filter_trend' not in self.df.columns:
-            self.insert_butter_filter()
-        
-        for i in windows:
-            self.df[f'butter_filter_trend_ma_{i}'] = (self.df['butter_filter_trend']).ewm(span=i).mean()
-            self.df[f'close_normed_butter_filter_trend_ma_{i}'] = self.df['close'] / self.df[f'butter_filter_trend_ma_{i}']
-
-            if f'butter_filter_trend_ma_{i}' not in self.numeric_features:
-                self.numeric_features += [f'butter_filter_trend_ma_{i}', f'close_normed_butter_filter_trend_ma_{i}']
-    
-    def insert_trend_deviation(self):
-        if 'butter_filter_trend' not in self.df.columns:
-            self.insert_butter_filter()
-        
-        self.df['butter_filter_trend_deviation'] = (self.df['close'] - self.df['butter_filter_trend']) / self.df['butter_filter_trend']
-        self.df['close_normed_butter_filter_trend_deviation'] = self.df['close'] / self.df['butter_filter_trend_deviation']
-
-        if 'butter_filter_trend_deviation' not in self.numeric_features:
-            self.numeric_features += ['butter_filter_trend_deviation', 'close_normed_butter_filter_trend_deviation']
-
             
     def insert_all(self, common_windows= None):
         if common_windows is None:
@@ -368,6 +202,7 @@ class FinData():
         self.insert_shifts_norms(common_windows)
         self.insert_time_features()
         self.insert_holidays()
+        self.insert_seasons()
         self.insert_rolling_means(common_windows)
         self.insert_exp_rolling_means(common_windows)
         self.insert_rsi(common_windows)
@@ -376,6 +211,9 @@ class FinData():
         self.insert_stochastic_oscillator(common_windows)
         self.insert_random_prediction()
         self.insert_butter_filter()
+        self.insert_trend_rsi()
+        self.insert_trend_rolling_means()
+        self.insert_trend_deviation()
 
     def get_columns(self):
         return self.df.columns
